pyproject.toml
[tool.poetry]
name = "cdm_unofficial_api_2"
version = "0.1.0"
description = ""
authors = ["Cleyson Lima <cleysonph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
python-decouple = "^3.3"
dj-database-url = "^0.5.0"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
